Move environment config into the mock impl itself.
